refactor(post): extract image upload logic to custom hook
- Refactored image upload logic in AddPost.jsx by extracting related states
and functions (image, imagePreview, uploading, uploadImage, getURL) into
a custom hook useImageUpload.
- This improves code reusability and reduces component size.
- Also returned setImage and setImagePreview for use in edit mode (fetchRecipe).
